fix: work well when submitting/updating an issue or showing its preview
