+++ manifest.json
@@ -1,8 +1,8 @@
 {
   "manifest_version": 2,
   "name": "PlantUML Visualizer",
-  "description": "A tool for visualizing PlantUML code.",
-  "version": "1.2.0",
+  "description": "A Chrome / Firefox extension for visualizing PlantUML descriptions.",
+  "version": "1.3.0",
   "browser_action": {
     "default_icon": "icon/icon16.png"
   },
