+++ src/encoder/PlantUmlEncoder.ts
@@ -1,5 +1,5 @@
 import { deflate } from 'zlib.es';
-import { Constants } from './Constants';
+import { Constants } from '../Constants';
 
 export const ImageSrcPrefix = `${Constants.imgSrcUrl}/svg/`;
